rp-chat-backend/requirements.txt
kafka-python
flask
flask-cors
python-dotenv